backend/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# File Handling
python-multipart==0.0.6
openpyxl==3.1.2
pandas==2.1.4
aiofiles==23.2.1
pillow==10.2.0
python-magic==0.4.27

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# AI Chat
groq==0.4.2

# Configuration
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Development (optional, comment out for production)
# pytest==7.4.4
# pytest-asyncio==0.23.3
# httpx==0.26.0
